split modals to components
